Refactor error handling and route setup to enforce Send trait for RunnerExecutor; update dependencies and improve resource initialization
