+++ Cargo.toml
@@ -54,7 +54,7 @@ headers = { version = "0.4", default-features = false }
 hex = { version = "0.4", default-features = false }
 http = { version = "1", default-features = false }
 httparse = { version = "1", default-features = false }
-ipinfo = { git = "https://github.com/rakshith-ravi/ipinfo-rust", branch = "feature/upgrade-reqwest", default-features = false }
+ipinfo = { version = "3", default-features = false }
 ipnetwork = { version = "0.20", default-features = false }
 jsonwebtoken = { version = "9", default-features = false }
 k8s-openapi = { version = "0.24", default-features = false }
