+++ runners/common/src/routes/auth/mod.rs
@@ -12,7 +12,7 @@ use self::{login::*, sign_up::*};
 #[instrument(skip(state))]
 pub async fn setup_routes<E>(state: &AppState<E>) -> Router
 where
-	E: RunnerExecutor + Clone + 'static,
+	E: RunnerExecutor + Send + 'static,
 {
 	Router::new()
 		.mount_endpoint(login, state)
