+++ runners/common/src/routes/workspace/mod.rs
@@ -8,7 +8,7 @@ pub mod deployment;
 #[instrument(skip(state))]
 pub async fn setup_routes<E>(state: &AppState<E>) -> Router
 where
-	E: RunnerExecutor + Clone + 'static,
+	E: RunnerExecutor + Send + 'static,
 {
 	Router::new().merge(deployment::setup_routes(state).await)
 }
